Add operations tab, fundamental indicators, and backtest improvements
- New Operações tab: real-time trade signals with confirm/cancel, per-asset
  position tracking, pending buybacks, automatic hourly background checker
  (AppConfig.ready + daemon thread)
- Fundamental indicators via investidor10.com.br scraping (P/L, P/VP, DY,
  ROE, LPA, VPA, EV/EBITDA, Margem Líquida, Dív/Patrim) with buy/sell/wait
  recommendation scoring
- Backtest: added min_hold_pct (always keep ≥X% of shares), B3 fee (0.0305%),
  removed redundant max_pct parameter, unified trailing-stop threshold
- Status page: fix decimal formatting to 2 places with Brazilian locale
- Template fix: {% load app_filters %} after {% extends %} to avoid tag errors

+++ monitor/views.py
@@ -6,7 +6,7 @@ from market import get_current_price, is_market_open
 
 def _stock_status(s):
     avg = float(s.avg_price)
-    qty = float(s.quantity)
+    qty = int(s.quantity)
     min_pct = float(s.min_profit_pct)
     max_pct = float(s.max_profit_pct)
